Change comboBox KursID to textfield and make new field Kursbezeichnung; add new Logic for these changes

+++ Hochschuldatenbank/Hochschulverwaltung.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using System.Windows.Forms.VisualStyles;
 using Accessibility;
 using static System.Windows.Forms.VisualStyles.VisualStyleElement;
 
@@ -20,22 +21,38 @@ namespace Hochschuldatenbank
     public partial class Hochschulverwaltung : Form
     {
         // Properties and Fields
-        private DB DataBase { get; } 
+        private DB DataBase { get; }
         private SqlConnection connection { get; }
         private int lastSelectedKursID = 0;
         private int lastSelectedMatrNr = 0;
         private int lastSelectedSemester = 0;
+        private int lastSelectedPersNr = 0;
 
-        public Hochschulverwaltung() // constructor
+        /// <summary>
+        /// This is the constructor of this class
+        /// </summary>
+        public Hochschulverwaltung() 
         {
             DataBase = new DB();
             connection = DataBase.DataBaseConnection();
             InitializeComponent();
+            InitializeCombobox();
             // fill comboBoxes with data
             FillComboData();
-            // ClearAllFieldsVerwaltung();
             // Show DatGridView
             ShowVerwaltung();
+
+        }
+
+        /// <summary>
+        /// This method initialize method comboBox KursBezeichnung.
+        /// If content in comboBox changes through user, method will
+        /// call
+        /// </summary>
+        void InitializeCombobox()
+        {
+            comboBoxVerwaltungKursBez.SelectedText += comboBoxVerwaltungKursBez_SelectedIndexChanged;
+
         }
 
         /// <summary>
@@ -82,7 +99,8 @@ namespace Hochschuldatenbank
         private void btnVerwaltungSpeichern_Click(object sender, EventArgs e)
         {
             // check, if there is data in comboBoxes on Screen
-            if (comboBoxVerwaltungKursID.Text == string.Empty ||
+            if (comboBoxVerwaltungKursBez.Text == string.Empty ||
+                txtBoxVerwaltungKursID.Text == string.Empty ||
                 comboBoxVerwaltungMatrNr.Text == string.Empty ||
                 comboBoxVerwaltungPersNr.Text == string.Empty ||
                 comboBoxVerwaltungSemester.Text == string.Empty)
@@ -92,20 +110,22 @@ namespace Hochschuldatenbank
             }
 
             // Parse Text (string) explicitly to integer
-            int verwaltungKursID = int.Parse(comboBoxVerwaltungKursID.Text);
+            int verwaltungKursID = int.Parse(txtBoxVerwaltungKursID.Text);
             int verwaltungMatrNr = int.Parse(comboBoxVerwaltungMatrNr.Text);
             int verwaltungPersNr = int.Parse(comboBoxVerwaltungPersNr.Text);
             int verwaltungSemester = int.Parse(comboBoxVerwaltungSemester.Text);
+            // parsing not necessary for text data
+            string verwaltungKursBez = comboBoxVerwaltungKursBez.Text;
 
             // SQL-Query: fill data in table Dozenten
-            string query = string.Format("INSERT INTO KursTeilnahmen VALUES('{0}', '{1}', '{2}', '{3}');",
-                verwaltungKursID, verwaltungMatrNr, verwaltungSemester, verwaltungPersNr);
+            string query = string.Format("INSERT INTO KursTeilnahmen VALUES('{0}', '{1}', '{2}', '{3}', '{4}');",
+                verwaltungKursID, verwaltungKursBez, verwaltungMatrNr, verwaltungSemester, verwaltungPersNr);
 
             // Ececute Query
             DataBase.DataBaseExecuteQuery(query, connection);
-            
+
             //ClearAllFieldsVerwaltung();
-            
+
             // Show actual data
             ShowVerwaltung();
         }
@@ -120,14 +140,16 @@ namespace Hochschuldatenbank
             // check, if there is row selected on Screen in DataGridView
             if (lastSelectedKursID != 0 &&
                 lastSelectedMatrNr != 0 &&
-                lastSelectedSemester != 0)
+                lastSelectedSemester != 0 &&
+                lastSelectedPersNr != 0)
             {
                 // SQL-Query: delete data in SQL database in tbale KursTeilnahmen
                 string query = string.Format("DELETE FROM KursTeilnahmen WHERE KursID = {0}" +
                                              "AND Matrikelnummer = {1}" +
-                                             "AND Semester = {2};",
-                                             lastSelectedKursID, lastSelectedMatrNr, lastSelectedSemester);
-                
+                                             "AND Semester = {2} AND Personalnummer = {3};",
+                                             lastSelectedKursID, lastSelectedMatrNr, 
+                                             lastSelectedSemester, lastSelectedPersNr);
+
                 // Execute query
                 DataBase.DataBaseExecuteQuery(query, connection);
             }
@@ -137,8 +159,6 @@ namespace Hochschuldatenbank
                 return;
             }
 
-            // ClearAllFieldsVerwaltung();
-            
             // Show actual data
             ShowVerwaltung();
         }
@@ -151,15 +171,17 @@ namespace Hochschuldatenbank
         private void GridViewVerwaltung_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
             // choose row in DataGridView and display these data as text in GridView
-            comboBoxVerwaltungKursID.Text = GridViewVerwaltung.SelectedRows[0].Cells[0].Value.ToString();
-            comboBoxVerwaltungMatrNr.Text = GridViewVerwaltung.SelectedRows[0].Cells[1].Value.ToString();
-            comboBoxVerwaltungSemester.Text = GridViewVerwaltung.SelectedRows[0].Cells[2].Value.ToString();
-            comboBoxVerwaltungPersNr.Text = GridViewVerwaltung.SelectedRows[0].Cells[3].Value.ToString();
+            comboBoxVerwaltungKursBez.Text = GridViewVerwaltung.SelectedRows[0].Cells[1].Value.ToString();
+            txtBoxVerwaltungKursID.Text = GridViewVerwaltung.SelectedRows[0].Cells[0].Value.ToString();
+            comboBoxVerwaltungMatrNr.Text = GridViewVerwaltung.SelectedRows[0].Cells[2].Value.ToString();
+            comboBoxVerwaltungSemester.Text = GridViewVerwaltung.SelectedRows[0].Cells[3].Value.ToString();
+            comboBoxVerwaltungPersNr.Text = GridViewVerwaltung.SelectedRows[0].Cells[4].Value.ToString();
 
             // Parse text data (string) into integer
-            lastSelectedKursID = int.Parse(comboBoxVerwaltungKursID.Text);
+            lastSelectedKursID = int.Parse(txtBoxVerwaltungKursID.Text);
             lastSelectedMatrNr = int.Parse(comboBoxVerwaltungMatrNr.Text);
             lastSelectedSemester = int.Parse(comboBoxVerwaltungSemester.Text);
+            lastSelectedPersNr = int.Parse(comboBoxVerwaltungPersNr.Text);
         }
 
         /// <summary>
@@ -180,10 +202,10 @@ namespace Hochschuldatenbank
         {
             // initialize List with tables and empty Lists to fill with appropriate data
             var tableNames = new List<string> { "Dozenten", "Kurse", "Studierende" };
-            List<string> comboDataPersonalnummer = new List<string>();
-            List<string> comboDataKursID = new List<string>();
-            List<string> comboDataMatrikelnummer = new List<string>();
-            List<string> comboDataSemester = new List<string>();
+            var comboDataPersonalnummer = new List<string>();
+            var comboDataKursBezeichnung = new List<string>();
+            var comboDataMatrikelnummer = new List<string>();
+            var comboDataSemester = new List<string>();
 
             foreach (var tableName in tableNames)
             {
@@ -199,7 +221,7 @@ namespace Hochschuldatenbank
                     {
                         // check which table is currently in loop
                         switch (tableName)
-                        {   
+                        {
                             // table Dozenten
                             case "Dozenten":
                                 // write data from table Dozenten, column Personalnummer in appropriate List
@@ -207,10 +229,16 @@ namespace Hochschuldatenbank
                                 break;
                             // table Kurse
                             case "Kurse":
-                                // write data from table Kurse, column KursID in appropriate List
-                                comboDataKursID.Add(row["KursID"].ToString());
-                                // write data from table Kurse, column Semester in appropriate List
-                                comboDataSemester.Add(row["Semester"].ToString());
+                                // write data from table Kurse, column Semester in appropriate List if not in List (no doubled entrees)
+                                if (!comboDataSemester.Contains(row["Semester"].ToString()))
+                                {
+                                    comboDataSemester.Add(row["Semester"].ToString());
+                                }
+                                // write data from table Kurse, column Kursbezeichnung in appropriate List if not in List (no doubled entrees)
+                                if (!comboDataKursBezeichnung.Contains(row["Kursbezeichnung"].ToString()))
+                                {
+                                    comboDataKursBezeichnung.Add(row["Kursbezeichnung"].ToString());
+                                }
                                 break;
                             // table Studierende
                             case "Studierende":
@@ -226,8 +254,8 @@ namespace Hochschuldatenbank
                 }
             }
             // fill actual data from lists in appropriate ComboBoxes
+            comboBoxVerwaltungKursBez.DataSource = comboDataKursBezeichnung;
             comboBoxVerwaltungPersNr.DataSource = comboDataPersonalnummer;
-            comboBoxVerwaltungKursID.DataSource = comboDataKursID;
             comboBoxVerwaltungSemester.DataSource = comboDataSemester;
             comboBoxVerwaltungMatrNr.DataSource = comboDataMatrikelnummer;
         }
@@ -236,14 +264,6 @@ namespace Hochschuldatenbank
         /// This method deletes all Text information in ComboBoxes on Screen.
         /// Currently not used!
         /// </summary>
-        private void ClearAllFieldsVerwaltung()
-        {   
-            // write an empty string into ComboBoxes
-            comboBoxVerwaltungKursID.Text = string.Empty;
-            comboBoxVerwaltungMatrNr.Text = string.Empty;
-            comboBoxVerwaltungSemester.Text = string.Empty;
-            comboBoxVerwaltungPersNr.Text = string.Empty;
-        }
 
         /// <summary>
         /// This method provides updating data in database.
@@ -253,21 +273,50 @@ namespace Hochschuldatenbank
         private void btnVerwaltungBearbeiten_Click(object sender, EventArgs e)
         {
             // Parse Text data from ComboBoxes to integer values
-            int kursId = int.Parse(comboBoxVerwaltungKursID.Text);
+            int kursId = int.Parse(txtBoxVerwaltungKursID.Text);
             int matrNr = int.Parse(comboBoxVerwaltungMatrNr.Text);
             int semester = int.Parse(comboBoxVerwaltungSemester.Text);
             int persNr = int.Parse(comboBoxVerwaltungPersNr.Text);
+            
+            // Parsing Text data in textfield not necessary
+            string kursBez = comboBoxVerwaltungKursBez.Text;
 
             // SQL-Query for Updating data in database
             string query = string.Format("UPDATE KursTeilnahmen SET KursID='{0}'," +
-                                         "Matrikelnummer='{1}', Semester='{2}'," +
-                                         "Personalnummer='{3}' WHERE KursID = {4}" +
-                                         "AND Matrikelnummer = {5};",
-                                         kursId, matrNr, semester, persNr,
+                                         "Kursbezeichnung='{1}', Matrikelnummer='{2}', " +
+                                         "Semester='{3}', Personalnummer='{4}'" +
+                                         "WHERE KursID = {5} AND Matrikelnummer = {6};",
+                                         kursId, kursBez, matrNr, semester, persNr,
                                          lastSelectedKursID, lastSelectedMatrNr);
 
             // Execute SQL-Query
             DataBase.DataBaseExecuteQuery(query, connection);
+            // Show actual data
+            ShowVerwaltung();
+        }
+
+        /// <summary>
+        /// This method is for textfield KursID and should change content,
+        /// if user changes comboBox Kursbezeichnung
+        /// </summary>
+        /// <param name="sender">source of incoming data/event</param>
+        /// <param name="e">Eventhandler</param>
+        private void comboBoxVerwaltungKursBez_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            string selectedKursBez = comboBoxVerwaltungKursBez.Text;
+            // SQL-Query
+            string query = string.Format("SELECT * FROM Kurse WHERE Kursbezeichnung = '{0}';", selectedKursBez);
+            // Execute SQL-Query
+            var dataTable = DataBase.DataBaseShowData(query, connection).Tables[0];
+
+            if (dataTable.Rows.Count > 0)
+            {
+                // Annahme: Die Ergebnisse befinden sich in der ersten Zeile der Tabelle
+                string kursID = dataTable.Rows[0]["KursID"].ToString();
+                // Show result in textfield Kursbezeichnung on Screen
+                txtBoxVerwaltungKursID.Text = kursID;
+            }
+
             // Show actual data
             ShowVerwaltung();
         }
